# [Performance] Benchmark AddCounter with FakeStatsPlugin and OTelPlugin