Sample output:

```
➜  grpc git:(otel-metrics-benchmark) ✗ bazel-bin/test/cpp/microbenchmarks/bm_stats_plugin                                                                      
WARNING: All log messages before absl::InitializeLog() is called are written to STDERR
I0000 00:00:1721695619.615582 2126186 config.cc:257] gRPC experiments enabled: call_status_override_on_cancellation, call_tracer_in_transport, event_engine_dns, event_engine_listener, monitoring_experiment, pick_first_new, trace_record_callops, work_serializer_clears_time_cache
2024-07-22T17:46:59-07:00                                    
Running bazel-bin/test/cpp/microbenchmarks/bm_stats_plugin
Run on (48 X 2450 MHz CPU s)            
CPU Caches:                                        
  L1 Data 32 KiB (x24)                                     
  L1 Instruction 32 KiB (x24)                     
  L2 Unified 512 KiB (x24)                                     
  L3 Unified 32768 KiB (x3)     
Load Average: 1.16, 0.85, 0.85
***WARNING*** Library was built as DEBUG. Timings may be affected.
---------------------------------------------------------------------------
Benchmark                                 Time             CPU   Iterations
---------------------------------------------------------------------------
BM_AddCounterWithFakeStatsPlugin       1738 ns         1738 ns       404265
BM_AddCounterWithOTelPlugin             757 ns          757 ns       928142
I0000 00:00:1721695621.304593 2126186 test_config.cc:186] TestEnvironment ends
```


<!--

If you know who should review your pull request, please assign it to that
person, otherwise the pull request would get assigned randomly.

If your pull request is for a specific language, please add the appropriate
lang label.

-->